requirements.txt
streamlit==1.31.0
numpy==1.24.3
Pillow==10.0.0
opencv-python-headless==4.8.0.74
scikit-learn==1.3.0

flask==2.3.2
